DropIn Checkout Implementation moved to PresentCheckoutFragment.kt
