+++ app/src/main/java/io/inai/android_sample_integration/headless/HeadlessFlowsFragment.kt
@@ -4,19 +4,12 @@ import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.View
 import androidx.navigation.fragment.findNavController
+import io.inai.android_sample_integration.MainActivity.Companion.ARG_PAYMENT_OPERATION
+import io.inai.android_sample_integration.PaymentOperation
 import io.inai.android_sample_integration.R
 import kotlinx.android.synthetic.main.fragment_headless.*
 
-enum class HeadlessOperation {
-    MakePayment, PayWithSavedPaymentMethod
-}
-
-
 class HeadlessFragment : Fragment(R.layout.fragment_headless) {
-
-    companion object {
-        const val ARG_HEADLESS_OPERATION = "arg_payment_option"
-    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -24,7 +17,7 @@ class HeadlessFragment : Fragment(R.layout.fragment_headless) {
         btn_make_payment.setOnClickListener {
             findNavController().navigate(
                 R.id.action_headlessFragment_to_checkoutFragment,
-                Bundle().apply { putSerializable(ARG_HEADLESS_OPERATION, HeadlessOperation.MakePayment) }
+                Bundle().apply { putSerializable(ARG_PAYMENT_OPERATION, PaymentOperation.MakePayment) }
             )
         }
 
@@ -35,7 +28,7 @@ class HeadlessFragment : Fragment(R.layout.fragment_headless) {
         btn_saved_payment_method.setOnClickListener {
             findNavController().navigate(
                 R.id.action_headlessFragment_to_checkoutFragment,
-                Bundle().apply { putSerializable(ARG_HEADLESS_OPERATION, HeadlessOperation.PayWithSavedPaymentMethod) }
+                Bundle().apply { putSerializable(ARG_PAYMENT_OPERATION, PaymentOperation.PayWithSavedPaymentMethod) }
             )
         }
