+++ app/src/main/res/layout/activity_drop_in.xml
@@ -6,12 +6,12 @@
     android:layout_height="match_parent"
     tools:context=".drop_in.DropInActivity">
 
-    <ProgressBar
-        android:id="@+id/progressBar"
-        style="?android:attr/progressBarStyle"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerInParent="true"
-        android:visibility="invisible" />
+    <androidx.fragment.app.FragmentContainerView
+        android:id="@+id/dropin_host_fragment"
+        android:name="androidx.navigation.fragment.NavHostFragment"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        app:defaultNavHost="true"
+         />
 
 </RelativeLayout>
